requirements.txt
altair==4.2.0
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
cachetools==5.2.0
category-encoders==2.5.1.post0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorcet==3.0.1
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
holoviews==1.15.3
hvplot==0.8.2
hyperopt==0.2.7
idna==3.4
imbalanced-learn==0.10.0
imblearn==0.0
importlib-metadata==5.2.0
ipykernel==6.19.4
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
packaging==22.0
pandas==1.5.2
panel==0.14.2
param==1.12.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==10.0.1
pyct==0.4.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.16.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
xgboost==1.7.2
zipp==3.11.0